% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{calc_distance}
\alias{calc_distance}
\title{Correlation calculation based on rolling window with overlapping observations.}
\usage{
calc_distance(lstx, finddataleaksout, h)
}
\arguments{
\item{lstx}{list of time series}

\item{finddataleaksout}{list, the output generated from find_dataleaks function}

\item{h}{length of the window size}
}
\value{
matrix visualizing the output
}
\description{
Correlation calculation based on rolling window with overlapping observations.
}
