% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_dataleaks.R
\name{viz_dataleaks}
\alias{viz_dataleaks}
\title{Correlation calculation based on rolling window with overlapping observations.}
\usage{
viz_dataleaks(finddataleaksout)
}
\arguments{
\item{finddataleaksout}{list, the output generated from find_dataleaks function}
}
\value{
matrix visualizing the output
}
\description{
Correlation calculation based on rolling window with overlapping observations.
}
