% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_details.R
\name{viz_details}
\alias{viz_details}
\title{Correlation calculation based on rolling window with overlapping observations.}
\usage{
viz_details(finddataleaksout, reasondataleaksout)
}
\arguments{
\item{finddataleaksout}{list, the output generated from find_dataleaks function}

\item{reasondataleaksout}{dataframe, output generated from reason_dataleaks function}
}
\value{
matrix visualizing the output
}
\description{
Correlation calculation based on rolling window with overlapping observations.
}
