% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dataleaks.R
\name{find_dataleaks}
\alias{find_dataleaks}
\title{Correlation calculation based on rolling window with overlapping observations.}
\usage{
find_dataleaks(lstx, h, cutoff = 1)
}
\arguments{
\item{lstx}{list of time series}

\item{h}{length of forecast horizon}

\item{cutoff}{benchmark value for corr, default 1}
}
\value{
list of matching quantities
}
\description{
Correlation calculation based on rolling window with overlapping observations.
}
